# internal/interfaces/rest/openapi_types.gen.go
// Package rest provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.8.2 DO NOT EDIT.
package rest

import (
	openapi_types "github.com/deepmap/oapi-codegen/pkg/types"
)

const (
	BearerAuthScopes = "bearerAuth.Scopes"
)

// CreateUserRequest defines model for CreateUserRequest.
type CreateUserRequest struct {
	EmailAddress openapi_types.Email `json:"emailAddress"`
	Id           string              `json:"id"`
	Password     string              `json:"password"`
}

// Error defines model for Error.
type Error struct {
	Message string `json:"message"`
	Slug    string `json:"slug"`
}

// User defines model for User.
type User struct {
	EmailAddress string `json:"emailAddress"`
	Id           string `json:"id"`
}

// Users defines model for Users.
type Users struct {
	Users []User `json:"users"`
}

// CreateUserJSONBody defines parameters for CreateUser.
type CreateUserJSONBody CreateUserRequest

// CreateUserJSONRequestBody defines body for CreateUser for application/json ContentType.
type CreateUserJSONRequestBody CreateUserJSONBody